Implement basic error handling
In most places, we weren't handling errors. We were also using a custom
function that printed the errors and exited with an OS status of 1.

Considering this package will also be serving as a general purpose
library for inclusion in other Go programs (a Terraform provisioner, for
example), we can't include such functions within the library's core
functionality.

All error printing/exit status operations are now happening in main
where the CLI functionality is currently implemented. This may be
separated out into another package in the future.